backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
httpx
openai
alembic
python-multipart
email-validator
mercadopago
